Moved code back from the inline file.
